what i have so far

+++ bust_or_boom/nohrsc_plotting.py
@@ -2,30 +2,35 @@ import datetime
 import urllib.request
 import os
 
-import cartopy.crs as ccrs
-import cartopy.feature as cfeature
-import matplotlib.pyplot as plt
 import xarray
 
-# Remove any .nc files that already exist
-directory = next(os.walk("./"))
-for file_name in directory[-1]:
-    if file_name.endswith(".nc"):
-        os.remove(file_name)
 
-DATA_TIME = datetime.datetime(year=2022, month=1, day=7)
+DATA_TIME = datetime.datetime(year=2022, month=1, day=8)
 FILE_PATH = f"{DATA_TIME.strftime('%Y%m%d%H')}.nc"
 URL = (
         f"http://www.nohrsc.noaa.gov/snowfall_v2/data/"
         f"{DATA_TIME.strftime('%Y%m')}/sfav2_CONUS_24h_{DATA_TIME.strftime('%Y%m%d%H')}.nc"
     )
 
-urllib.request.urlretrieve(URL, FILE_PATH)
+if not os.path.exists(FILE_PATH):
+    # Remove any .nc files that already exist
+    directory = next(os.walk("./"))
+    for file_name in directory[-1]:
+        if file_name.endswith(".nc"):
+            os.remove(file_name)
 
-dataset = xarray.open_mfdataset(FILE_PATH)
+    urllib.request.urlretrieve(URL, FILE_PATH)
 
 
-def nohrsc_snow():
+def nohrsc_snow(extent):
+    min_lon, max_lon, min_lat, max_lat = extent
+
+    dataset = xarray.open_mfdataset(FILE_PATH)
+
+    mask_lon = (dataset.lon >= min_lon) & (dataset.lon <= max_lon)
+    mask_lat = (dataset.lat >= min_lat) & (dataset.lat <= max_lat)
+
+    dataset = dataset.where(mask_lon & mask_lat, drop=True)
     snow_data = dataset.Data * 39.3700787
 
     lats = snow_data.lat.values
